+++ src/test/java/ladder/domain/DirectionTest.java
@@ -1,10 +1,7 @@
-package ladder;
+package ladder.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import ladder.domain.Direction;
-import ladder.domain.RandomSingleton;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
