+++ src/test/java/ladder/domain/RewardsTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.Arrays;
 import java.util.List;
@@ -6,7 +6,6 @@ import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import ladder.domain.Rewards;
 import ladder.util.StringUtils;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
